+++ web/app/playground/_components/editor/index.tsx
@@ -25,6 +25,7 @@ import {
   transparentGridPatterns,
 } from "@/utils";
 import Caption from "./Caption";
+import EditorActionButtons from "./EditorActionButtons";
 
 interface EditorProps {
   code: string;
@@ -471,6 +472,25 @@ const Editor = React.forwardRef<HTMLDivElement, EditorProps>(
         >
           {/* Snip Area Wrapper - Window Controls + Code Content */}
           <div className="relative flex-1 flex flex-col min-h-0">
+            {/* Editor Action Buttons - Positioned outside and above code editor */}
+            {!isFullscreen && (
+              <div
+                data-export-ignore
+                className="absolute right-0 opacity-0 group-hover:opacity-100 transition-opacity duration-200"
+                style={{
+                  top: "-44px", // Position above the editor
+                  zIndex: 50,
+                }}
+              >
+                <EditorActionButtons
+                  code={code}
+                  onClear={() => onChange("")}
+                  settings={settings}
+                  editorRef={containerRef.current}
+                />
+              </div>
+            )}
+
             <div
               className="relative flex flex-col flex-1 min-h-0"
               style={{